Improve docs
- Add links to the Bluetooth Core spec where relevant.
- Move `WithIndicator` and `CmdError` to the `transport` and `cmd` modules, respectively.
- General consistency improvements
