+++ src/cmd/controller_baseband.rs
@@ -1,4 +1,4 @@
-//! Bluetooth Core Specification Vol 4, Part E, §7.3
+//! Controller & Baseband commands [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-5ced811b-a6ce-701a-16b2-70f2d9795c05)
 
 use crate::cmd;
 use crate::param::{
@@ -7,7 +7,7 @@ use crate::param::{
 };
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.1
+    /// Set Event Mask command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-9cf88217-77b4-aeb6-61fb-d1129d48a67c)
     SetEventMask(CONTROL_BASEBAND, 0x0001) {
         Params = EventMask;
         Return = ();
@@ -15,7 +15,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.2
+    /// Reset command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-b0aaafb1-0601-865c-2703-4f4caa4dee2e)
     Reset(CONTROL_BASEBAND, 0x0003) {
         Params = ();
         Return = ();
@@ -23,12 +23,11 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.35
+    /// Read Transmit Power Level command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-7205a3ee-15c7-cc48-c512-a959b4e3f560)
     ReadTransmitPowerLevel(CONTROL_BASEBAND, 0x002d) {
         ReadTransmitPowerLevelParams {
             kind: PowerLevelKind,
         }
-        /// Bluetooth Core Specification Vol 4, Part E, §7.3.35
         ReadTransmitPowerLevelReturn {
             tx_power_level: i8,
         }
@@ -37,7 +36,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.38
+    /// Set Controller To Host Flow Control command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-64d757fc-e1da-329f-6d6a-16453750f325)
     SetControllerToHostFlowControl(CONTROL_BASEBAND, 0x0031) {
         Params = ControllerToHostFlowControl;
         Return = ();
@@ -45,7 +44,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.39
+    /// Host Buffer Size command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-99748527-de87-bf9c-935d-baf7e3f35b12)
     HostBufferSize(CONTROL_BASEBAND, 0x0033) {
         HostBufferSizeParams {
             host_acl_data_packet_len: u16,
@@ -58,7 +57,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.40
+    /// Host Number Of Completed Packets command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-14569cb0-16b0-7dc0-a1d4-e5a4ef44d81a)
     ///
     /// *Note:* This command only returns a [`CommandComplete`](crate::event::CommandComplete) event on error. No event is generated on success.
     HostNumberOfCompletedPackets(CONTROL_BASEBAND, 0x0035)  {
@@ -68,7 +67,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.69
+    /// Set Event Mask Page 2 command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-4e91f200-b802-45d6-9282-fd03c0dfefbe)
     SetEventMaskPage2(CONTROL_BASEBAND, 0x0063) {
         Params = EventMaskPage2;
         Return = ();
@@ -76,10 +75,9 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.93
+    /// Read Authenticated Payload Timeout command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-36c940e9-a654-f07f-75cd-2cbcf2d6adf6)
     ReadAuthenticatedPayloadTimeout(CONTROL_BASEBAND, 0x007b) {
         Params = ConnHandle;
-        /// Bluetooth Core Specification Vol 4, Part E, §7.3.93
         ReadAuthenticatedPayloadTimeoutReturn {
             timeout: Duration<10_000>,
         }
@@ -88,7 +86,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.3.94
+    /// Write Authenticated Payload Timeout command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-9bb1bb66-1857-83d8-8954-677f773225f9)
     WriteAuthenticatedPayloadTimeout(CONTROL_BASEBAND, 0x007c) {
         WriteAuthenticatedPayloadTimeoutParams {
             timeout: Duration<10_000>,
