+++ src/cmd/info.rs
@@ -1,13 +1,12 @@
-//! Bluetooth Core Specification Vol 4, Part E, §7.4
+//! Informational parameters [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-42372304-c9ef-dcab-6905-4e5b64703d45)
 
 use super::cmd;
 use crate::param::{BdAddr, CmdMask, CoreSpecificationVersion, LmpFeatureMask};
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.4.1
+    /// Read Local Version Information command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-cf7fef88-faa4-fd2e-7c00-ab1ec7985a19)
     ReadLocalVersionInformation(INFO_PARAMS, 0x0001) {
         Params = ();
-        /// Bluetooth Core Specification Vol 4, Part E, §7.4.1
         ReadLocalVersionInformationReturn {
             hci_version: CoreSpecificationVersion,
             hci_subversion: u16,
@@ -19,7 +18,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.4.2
+    /// Read Local Supported Commands command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-d9df0f48-030f-0567-ecf3-8304df5c3eb0)
     ReadLocalSupportedCmds(INFO_PARAMS, 0x0002) {
         Params = ();
         Return = CmdMask;
@@ -27,7 +26,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.4.3
+    /// Read Local Supported Features command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-17c54ce9-1772-096f-c512-ba080bd11d04)
     ReadLocalSupportedFeatures(INFO_PARAMS, 0x0003) {
         Params = ();
         Return = LmpFeatureMask;
@@ -35,7 +34,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.4.6
+    /// Read BD_ADDR command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-151a8bec-71be-df54-2043-92d366376c53)
     ReadBdAddr(INFO_PARAMS, 0x0009) {
         Params = ();
         Return = BdAddr;
