+++ src/cmd/link_control.rs
@@ -1,10 +1,10 @@
-//! Bluetooth Core Specification Vol 4, Part E, §7.1
+//! Link Control commands [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-fe2a33d3-28f4-9fd1-4d08-62286985c05e)
 
 use crate::cmd;
 use crate::param::{ConnHandle, DisconnectReason};
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.1.6
+    /// Disconnect command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-6bb8119e-aa67-d517-db2a-7470c35fbf4a)
     Disconnect(LINK_CONTROL, 0x0006) {
         DisconnectParams {
             handle: ConnHandle,
@@ -15,7 +15,7 @@ cmd! {
 }
 
 cmd! {
-    /// Bluetooth Core Specification Vol 4, Part E, §7.1.23
+    /// Read Remote Version Information command [📖](https://www.bluetooth.com/wp-content/uploads/Files/Specification/HTML/Core-54/out/en/host-controller-interface/host-controller-interface-functional-specification.html#UUID-ebf3c9ac-0bfa-0ed0-c014-8f8691ea3fe5)
     ReadRemoteVersionInformation(LINK_CONTROL, 0x001d) {
         Params = ConnHandle;
     }
